backend/requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
reportlab==4.0.4
email-validator==2.1.0
fastapi-mail==1.4.1
